+++ projects/ids-enterprise-ng/src/lib/context-menu/soho-context-menu.directive.ts
@@ -1,20 +1,95 @@
 import {
   AfterViewInit,
+  Component,
   Directive,
   ElementRef,
   EventEmitter,
+  HostBinding,
   Input, NgZone,
   OnDestroy,
   Output
 } from '@angular/core';
 
+/**
+ * SUB COMPONENT: SOHO-CONTEXT-MENU-HEADING
+ */
+@Component({
+  selector: 'li[soho-context-menu-heading]', // tslint:disable-line
+  template: `<ng-content></ng-content>`
+})
+export class SohoContextMenuHeadingComponent {
+  @HostBinding('class.heading') get isHeading() { return true; }
+}
+
+/**
+ * SUB COMPONENT: SOHO-CONTEXT-MENU-SHORTCUT-TEXT
+ */
+@Component({
+  selector: 'span[soho-context-menu-shortcut-text]', // tslint:disable-line
+  template: `<ng-content></ng-content>`
+})
+export class SohoContextMenuShortCutTextComponent {
+  @HostBinding('class.shortcut-text') get isShortCutText() { return true; }
+}
+
+/**
+ * SUB COMPONENT: SOHO-CONTEXT-MENU-SEPARATOR
+ */
+@Component({
+  selector: 'li[soho-context-menu-separator]', // tslint:disable-line
+  template: `<ng-content></ng-content>`
+})
+export class SohoContextMenuSeparatorComponent {
+  @HostBinding('class.separator') get isSeparator() { return true; }
+  @HostBinding('class.single-selectable-section') @Input() singleSelectableSection = false;
+  @HostBinding('class.multi-selectable-section') @Input() multiSelectableSection = false;
+}
+
+/**
+ * SUB COMPONENT: SOHO-CONTEXT-MENU-LABEL
+ */
+@Component({
+  selector: 'a[soho-context-menu-label]', // tslint:disable-line
+  template: `<ng-content></ng-content>`
+})
+export class SohoContextMenuItemLabelComponent {
+  @HostBinding('attr.href') get hrefAttr() {
+    if (this.menuId) {
+      return '#' + this.menuId;
+    }
+    if (this.menuUrl) {
+      return this.menuUrl;
+    }
+    return '#';
+  }
+  @Input() menuId: string;
+  @Input() menuUrl: string;
+}
+
+/**
+ * SUB COMPONENT: SOHO-CONTEXT-MENU-ITEM
+ */
+@Component({
+  selector: 'li[soho-context-menu-item]', // tslint:disable-line
+  template: `<ng-content></ng-content>`
+})
+export class SohoContextMenuItemComponent {
+  @HostBinding('class.is-checked') @Input() isChecked: boolean;
+  @HostBinding('class.is-selectable') @Input() isSelectable = false;
+  @HostBinding('class.is-disabled') @Input() isDisabled = false;
+  @HostBinding('class.is-indented') @Input() isIndented = false;
+  @HostBinding('class.submenu') @Input() subMenu = false;
+}
+
+/**
+ * MAIN DIRECTIVE: SOHO-CONTEXT-MENU
+ */
 @Directive({
   selector: '[soho-context-menu]', // tslint:disable-line
 })
 
 export class SohoContextMenuDirective implements AfterViewInit, OnDestroy {
   private jQueryElement: JQuery;
-
   private contextMenu: SohoPopupMenuStatic;
 
   // -------------------------------------------
@@ -52,6 +127,93 @@ export class SohoContextMenuDirective implements AfterViewInit, OnDestroy {
     }
   }
 
+  /** Auto Focus */
+  @Input() set autoFocus(autoFocus: boolean) {
+    this.options.autoFocus = autoFocus;
+    if (this.contextMenu) {
+      this.contextMenu.settings.autoFocus = autoFocus;
+    }
+  }
+
+  get autoFocus(): boolean {
+    if (this.contextMenu) {
+      return this.contextMenu.settings.autoFocus;
+    }
+    return this.options.autoFocus;
+  }
+
+  /** Mouse focus. */
+  @Input() set mouseFocus(mouseFocus: boolean) {
+    this.options.mouseFocus = mouseFocus;
+    if (this.contextMenu) {
+      this.contextMenu.settings.mouseFocus = mouseFocus;
+    }
+  }
+
+  get mouseFocus(): boolean {
+    if (this.contextMenu) {
+      return this.contextMenu.settings.mouseFocus;
+    }
+    return this.options.mouseFocus;
+  }
+
+  /** Attach to body. */
+  @Input() set attachToBody(attachToBody: boolean) {
+    this.options.attachToBody = attachToBody;
+    if (this.contextMenu) {
+      this.contextMenu.settings.attachToBody = attachToBody;
+    }
+  }
+
+  get attachToBody(): boolean {
+    if (this.contextMenu) {
+      return this.contextMenu.settings.attachToBody;
+    }
+    return this.options.attachToBody;
+  }
+
+  @Input() set placementOpts(placementOpts: SohoPopupmenuPlacementOpts) {
+    this.options.placementOpts = placementOpts;
+    if (this.contextMenu) {
+      this.contextMenu.settings.placementOpts = placementOpts;
+    }
+  }
+
+  get placementOpts(): SohoPopupmenuPlacementOpts {
+    if (this.contextMenu) {
+      return this.contextMenu.settings.placementOpts;
+    }
+    return this.options.placementOpts;
+  }
+
+  @Input() set offset(offset: SohoPopupmenuOffset) {
+    this.options.offset = offset;
+    if (this.contextMenu) {
+      this.contextMenu.settings.offset = offset;
+    }
+  }
+
+  get offset(): SohoPopupmenuOffset {
+    if (this.contextMenu) {
+      return this.contextMenu.settings.offset;
+    }
+    return this.options.offset;
+  }
+
+  @Input() set removeOnDestroy(removeOnDestroy: boolean) {
+    this.options.removeOnDestroy = removeOnDestroy;
+    if (this.contextMenu) {
+      this.contextMenu.settings.removeOnDestroy = removeOnDestroy;
+    }
+  }
+
+  get removeOnDestroy(): boolean {
+    if (this.contextMenu) {
+      return this.contextMenu.settings.removeOnDestroy;
+    }
+    return this.options.removeOnDestroy;
+  }
+
   /** beforeOpen - ajax callback for open event */
   @Input() set beforeOpen(beforeOpen: SohoPopupMenuSourceFunction) {
     this.options.beforeOpen = beforeOpen;
@@ -64,7 +226,6 @@ export class SohoContextMenuDirective implements AfterViewInit, OnDestroy {
     if (this.contextMenu) {
       return this.contextMenu.settings.beforeOpen;
     }
-
     return this.options.beforeOpen;
   }