+++ src/app/app.routes.ts
@@ -38,6 +38,7 @@ import { ColumnGroupedXaxisTwolineDemoComponent } from './column-grouped/column-
 import { ColumnStackedDemoComponent } from './column-stacked/column-stacked.demo';
 import { CompletionChartDemoComponent } from './completion-chart/completion-chart.demo';
 import { ContextMenuDemoComponent } from './context-menu/context-menu.demo';
+import { ContextMenuToggleDemoComponent } from './context-menu/context-menu-toggle.demo';
 import { ContextualActionPanelDemoComponent } from './contextual-action-panel/contextual-action-panel.demo';
 import { DataGridAngularEditorDemoComponent } from './datagrid/datagrid-angular-editor.demo';
 import { DataGridAngularFormatterDemoComponent } from './datagrid/datagrid-angular-formatter.demo';
@@ -234,6 +235,7 @@ export const routes: Routes = [
   { path: 'column-stacked', component: ColumnStackedDemoComponent },
   { path: 'completion-chart', component: CompletionChartDemoComponent },
   { path: 'context-menu', component: ContextMenuDemoComponent },
+  { path: 'context-menu-toggle', component: ContextMenuToggleDemoComponent },
   { path: 'contextual-action-panel', component: ContextualActionPanelDemoComponent },
   { path: 'datagrid-breadcrumb', component: DataGridBreadcrumbDemoComponent },
   { path: 'datagrid-dirty-indication', component: DataGridDirtyIndicationDemoComponent },